requirements.txt
aiohttp==3.8.1
aioice==0.7.6
aiortc==1.3.0
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
av==8.1.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bidict==0.21.4
boto3==1.20.49
botocore==1.23.49
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
distro==1.6.0
dnspython==2.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.1.3
fastapi==0.73.0
frozenlist==1.3.0
google-crc32c==1.3.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.7.1
motor==2.5.1
multidict==6.0.2
netifaces==0.11.0
paramiko==2.9.2
passlib==1.7.4
pycparser==2.21
pydantic==1.9.0
pyee==9.0.4
pylibsrtp==0.7.1
pymongo==3.12.3
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-multipart==0.0.5
python-socketio==5.5.1
PyYAML==5.4.1
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.6
starlette==0.17.1
texttable==1.6.4
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.4
websocket-client==0.59.0
yarl==1.7.2
